Moving services interfaces (ports) to domain layer.
